+++ trunk/BasicTools/ConnectionSelection.h
@@ -55,15 +55,15 @@ namespace Tinkercell
 		void setAllStraight(QSemaphore*,int);
 
 	public slots:
-		double getControlPointX(int a0,int a1,int a2);
-		double getControlPointY(int a0,int a1,int a2);
-		void setControlPoint(int a0,int a1,int i,double a2,double a3);
-		void setCenterPoint(int a0,double a1,double a2);
-		double getCenterPointX(int);
-		double getCenterPointY(int);
-		void setStraight(int,int);
+		double getControlPointX(long a0,long a1,int a2);
+		double getControlPointY(long a0,long a1,int a2);
+		void setControlPoint(long a0,long a1,int i,double a2,double a3);
+		void setCenterPoint(long a0,double a1,double a2);
+		double getCenterPointX(long);
+		double getCenterPointY(long);
+		void setStraight(long,int);
 		void setAllStraight(int);
-		void setLineWidth(int,double,int);
+		void setLineWidth(long,double,int);
 	};
 
 	/*! \brief This class listens to selection events from the scene and responds by changing colors
@@ -165,15 +165,15 @@ namespace Tinkercell
 		//QMenu connectorsMenu;
 		void connectTCFunctions();
 		static ConnectionSelection_FToS fToS;
-		static double _getControlPointX(int ,int ,int );
-		static double _getControlPointY(int ,int ,int );
-		static void _setControlPoint(int ,int ,int i, double ,double );
-		static double _getCenterPointX(int);
-		static double _getCenterPointY(int);
-		static void _setCenterPoint(int,double,double);
-		static void _setStraight(int,int);
+		static double _getControlPointX(long ,long ,int );
+		static double _getControlPointY(long ,long ,int );
+		static void _setControlPoint(long ,long ,int i, double ,double );
+		static double _getCenterPointX(long);
+		static double _getCenterPointY(long);
+		static void _setCenterPoint(long,double,double);
+		static void _setStraight(long,int);
 		static void _setAllStraight(int);
-		static void _setLineWidth(int,double,int);
+		static void _setLineWidth(long,double,int);
 
 		class LineTypeChanged : public QUndoCommand
 		{
